refactoring and tweaking graph UI
tweaked the UI so now the y-axis is in bytes, but the labels on each bar in the histogram is formatted to so the data over that interval in MB or GB all the way up to YB. Got the handy function from https://stackoverflow.com/a/18650828.

Also cleaned up the names of the graphs.